+++ Source/CesiumRuntime/Private/CesiumPrimitiveMetadata.cpp
@@ -1,4 +1,4 @@
-// Copyright 2020-2023 CesiumGS, Inc. and Contributors
+// Copyright 2020-2024 CesiumGS, Inc. and Contributors
 
 #include "CesiumPrimitiveMetadata.h"
 #include "CesiumGltf/AccessorView.h"