+++ Source/CesiumRuntime/Private/CesiumPropertyTexture.cpp
@@ -1,4 +1,4 @@
-// Copyright 2020-2023 CesiumGS, Inc. and Contributors
+// Copyright 2020-2024 CesiumGS, Inc. and Contributors
 
 #include "CesiumPropertyTexture.h"
 #include "CesiumGltf/Model.h"