+++ Source/CesiumRuntime/Private/CesiumTextureResource.cpp
@@ -1,3 +1,5 @@
+// Copyright 2020-2024 CesiumGS, Inc. and Contributors
+
 #include "CesiumTextureResource.h"
 #include "RenderUtils.h"
 #include "Stats/Stats.h"